requirements.txt
# =======================================================================================
# 📄 Dosya Adı   : requirements.txt
# 📁 Konum       : guard_pc/requirements.txt
# 📌 Açıklama    : Guard PC uygulaması için gerekli Python paketleri
# =======================================================================================

# Temel GUI ve Görüntü İşleme
opencv-python==4.8.1.78
Pillow==10.1.0
numpy==1.24.3

# PyTorch - Model İşlemleri
torch==2.1.1
torchvision==0.16.1
ultralytics==8.0.206

# Firebase ve Google Cloud
firebase-admin==6.2.0
google-cloud-firestore==2.12.0
google-cloud-storage==2.10.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1

# HTTP ve API İşlemleri
requests==2.31.0
fastapi==0.104.1
uvicorn==0.24.0
websockets==12.0

# Çevre Değişkenleri ve Yapılandırma
python-dotenv==1.0.0

# Bildirim Servisleri
twilio==8.10.0
telegram-bot==0.7.0
python-telegram-bot==20.6

# Loglama ve Yardımcı Araçlar
loguru==0.7.2
schedule==1.2.0
psutil==5.9.6

# Görüntü ve Video Streaming
flask==3.0.0
flask-socketio==5.3.6
eventlet==0.33.3

# GUI için ek kütüphaneler (tkinter Python ile gelir)
customtkinter==5.2.0
tkinter-tooltip==2.1.0

# Veri İşleme
pandas==2.1.3
python-dateutil==2.8.2

# Güvenlik
cryptography==41.0.7
PyJWT==2.8.0